requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
psycopg2-binary==2.9.7
blinker==1.6.3
reportlab==4.0.4
python-docx==0.8.11
gunicorn==21.2.0
